space/requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
huggingface_hub
pyyaml
xgboost